lambdas/dashboard/requirements.txt
boto3==1.40.25
botocore==1.40.25
python-dotenv==1.1.1
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.10